Update css for footer and header
