+++ app/assets/stylesheets/footer.scss
@@ -1,10 +1,15 @@
+.footer {
+  height: 160px;
+  width: 100%;
+}
+
 .footer-nav {
   // border-top: 1px solid #d8d8d8;
-  margin-top: 20px;
-  height: 160px;
+  height: 100%;
   display: flex;
   flex-direction: row;
   justify-content: space-around;
+  background: #222831;
   section {
     margin: 0;
     padding: 0;
@@ -12,35 +17,9 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    background-color: white;
-    border: 1px solid #d8d8d8;
+    color: white;
     width: 24.5%;
 
-  }
-}
-
-.footer{
-  flex: 4;
-  background-color: #222831;
-}
-
-
-
-.examples {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  h4 {
-    padding: 0px;
-    margin: 5px 0px;
-  }
-  button {
-    margin: 5px 0px;
-    height: 30px;
-    font-size: 14px;
-    width: 120px;
-    padding: 0px;
-    align-self: center;
   }
 }
 
@@ -57,7 +36,6 @@
     flex-direction: row;
     justify-content: space-between;
     align-items: center;
-    // border: 1px solid black;
   }
 }
 
@@ -65,7 +43,10 @@
   font-size: 1.7em;
   line-height: 0.9em;
   border: 0;
-  color: black;
+  color: white;
   padding: 0 5px;
-
+  &:hover {
+    background: transparent;
+    color: #ebebeb;
+  }
 }
